updating the server file structure to have models and controllers folders
